refactored

+++ src/main/java/com/alokhin/autoservice/security/config/SecurityConfig.java
@@ -49,7 +49,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(AuthenticationManagerBuilder auth) throws Exception {
         auth.userDetailsService(userDetailsService)
-            .passwordEncoder(new ShaPasswordEncoder(encodingStrength));
+            .passwordEncoder(new ShaPasswordEncoder(encodingStrength.intValue()));
     }
 
     @Override
